Renamed Mapper.php to SaveReader.php which only purpose now is to write the db.db file
